melts\fix(back): #3964.1 exclude git directory
Former-commit-id: ee6080a32a8e7363bdc5d375b9a14a712a139ff8

+++ melts/toolbox/drills/pull_repos.py
@@ -104,6 +104,8 @@ def delete_out_of_scope_files(group: str) -> bool:
         for path in utils.file.iter_rel_paths(path_to_repo):
             if match_file(spec_ignore.patterns, path):
                 path = os.path.join(path_to_fusion, repo_name, path)
+                if '.git/' in path:
+                    continue
                 if os.path.isfile(path):
                     os.unlink(path)
                 elif os.path.isdir(path):
